Simplify nodeToElement
